requirements.txt
websocket-client
requests
tls-client
pypasser
names
colorama
curl_cffi
rich
pydantic
pyyaml
typer[all]